move gmx_top.py into fpdb/fpdb for one-time installation
